0003 - update vehicle service

+++ Application/Interfaces/IVehicleService.cs
@@ -5,9 +5,9 @@ namespace Application.Interfaces
     public interface IVehicleService
     {
         Task<IEnumerable<VehicleDto>> GetAllAsync();
-        Task<VehicleDto?> GetByIdAsync(double id);
+        Task<VehicleDto?> GetByIdAsync(long id);
         Task AddAsync(CreateVehicleDto dto);
         Task UpdateAsync(UpdateVehicleDto dto);
-        Task DeleteAsync(double id);
+        Task DeleteAsync(long id);
     }
 }
